+++ src/Archive/Test.js
@@ -1,0 +1,18 @@
+// import React from "react";
+// import IngredientsDND from "./IngredientsDND";
+// import PlannerDND from "./PlannerDND";
+
+// const Test = () => {
+//   return (
+//     <div className="Test">
+//       <h1>Welcome to Joy's Meal Planner</h1>
+//       <h3>Hello from Test</h3>
+//       <div className="dashboard">
+//         <IngredientsDND />
+//         <PlannerDND />
+//       </div>
+//     </div>
+//   );
+// };
+
+// export default Test;
