+++ src/pages/Home.js
@@ -5,7 +5,6 @@ import DashboardDND from "../components/DND/DashboardDND";
 const Home = () => {
   return (
     <div className="Home">
-      <h1>Welcome to Joy's Meal Planner v2</h1>
       <IngredientAjax />
       <DashboardDND />
     </div>
