+++ src/pages/Main.js
@@ -3,7 +3,6 @@ import { Route, Switch } from "react-router-dom";
 import Home from "./Home";
 import Info from "./Info";
 import Planner from "./Planner";
-import Test from "./Test";
 
 const Main = () => {
   return (
@@ -11,9 +10,7 @@ const Main = () => {
       <Switch>
         <Route exact path="/" component={Home} />
         <Route exact path="/info" component={Info} />
-        {/* <Route path="/receipes/:ingredient" component={Recipes} /> */}
         <Route path="/planner" component={Planner} />
-        <Route path="/test" component={Test} />
       </Switch>
     </main>
   );
